Create README.md
Everyone has bad days once in a while, and sometimes, all it takes is a kind or supportive word to help you snap out of the funk. A compliment, a nice gesture, a smile or even an inspirational quote can brighten even the darkest of days. Take a look at these quotes that’ll cheer you up when you’re feeling blue. You can send them along to a friend who’s going through tough times when he or she needs some words of encouragement.

On Perspective
“There are only two ways to live your life. One is as though nothing is a miracle. The other is as though everything is a miracle.”
On Not Giving Up
“I haven’t failed. I just know 10,000 ways that won’t work.
On Keeping Your Chin Up
“Keep your face always toward the sunshine, and shadows will fall behind you.”
On Life’s Pace
“You are never too old to set another goal or to dream a new dream.
On Letting Go of the Past
“It isn’t where you came from. It’s where you’re going that counts.
 
On Starting Small
“If I cannot do great things, I can do small things in a great way.
On Other People’s Opinions
“When you do something beautiful and nobody noticed, do not be sad. For the sun every morning is a beautiful spectacle, and yet most of the audience still sleeps.
On Confidence
“Believe you can, and you’re halfway there.” 
On Positivity
“If you have good thoughts they will shine out of your face like sunbeams, and you will always look lovely.” 
On Taking Life as It Comes
“Life has got all those twists and turns. You’ve got to hold on tight and off you go